backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
databricks-sql-connector>=3.0.0
databricks-sdk>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.0
